+++ test/SolrExpress.Solr4.UnitTests/Search/Parameter/DefaultFieldParameterTests.cs
@@ -33,7 +33,7 @@ namespace SolrExpress.Solr4.UnitTests.Search.Parameter
             ((ISearchItemExecution<List<string>>)parameter).AddResultInContainer(container);
 
             // Assert
-            Assert.Equal(1, container.Count);
+            Assert.Single(container);
             Assert.Equal("df=id", container[0]);
         }
     }