+++ docs/about/release-notes.md
@@ -5,6 +5,7 @@ Notes about releases
 ## [5.x] - 2018-06-06
 
 ### Bug fix
+* Fix DI exceptions in .Net 4.7.1 ([#250](https://github.com/solr-express/solr-express/issues/250))
 
 ### Enhancement
 * Support to .Net 4.7.1 ([#248](https://github.com/solr-express/solr-express/issues/248))